requirements.txt
crewai
crewai-tools
langchain-groq
python-dotenv
rich
streamlit
pdfplumber